code organization and bug fixing
